kusto: refactoring the identity helpers
