+++ azurerm/internal/services/kusto/kusto_cluster_resource.go
@@ -50,7 +50,7 @@ func resourceArmKustoCluster() *schema.Resource {
 
 			"location": azure.SchemaLocation(),
 
-			"identity": azure.SchemaKustoIdentity(),
+			"identity": schemaIdentity(),
 
 			"sku": {
 				Type:     schema.TypeList,
@@ -259,7 +259,7 @@ func resourceArmKustoClusterCreateUpdate(d *schema.ResourceData, meta interface{
 	}
 
 	if v, ok := d.GetOk("trusted_external_tenants"); ok {
-		trustedExternalTenants := azure.ExpandKustoClusterTrustedExternalTenants(v.([]interface{}))
+		trustedExternalTenants := expandTrustedExternalTenants(v.([]interface{}))
 		clusterProperties.TrustedExternalTenants = trustedExternalTenants
 	}
 
@@ -276,7 +276,7 @@ func resourceArmKustoClusterCreateUpdate(d *schema.ResourceData, meta interface{
 
 	if _, ok := d.GetOk("identity"); ok {
 		kustoIdentityRaw := d.Get("identity").([]interface{})
-		kustoIdentity := azure.ExpandKustoIdentity(kustoIdentityRaw)
+		kustoIdentity := expandIdentity(kustoIdentityRaw)
 		kustoCluster.Identity = kustoIdentity
 	}
 
@@ -370,7 +370,7 @@ func resourceArmKustoClusterRead(d *schema.ResourceData, meta interface{}) error
 		d.Set("location", azure.NormalizeLocation(*location))
 	}
 
-	if err := d.Set("identity", azure.FlattenKustoIdentity(clusterResponse.Identity)); err != nil {
+	if err := d.Set("identity", flattenIdentity(clusterResponse.Identity)); err != nil {
 		return fmt.Errorf("Error setting `identity`: %s", err)
 	}
 
@@ -386,7 +386,7 @@ func resourceArmKustoClusterRead(d *schema.ResourceData, meta interface{}) error
 	}
 
 	if clusterProperties := clusterResponse.ClusterProperties; clusterProperties != nil {
-		d.Set("trusted_external_tenants", azure.FlattenKustoClusterTrustedExternalTenants(clusterProperties.TrustedExternalTenants))
+		d.Set("trusted_external_tenants", flattenTrustedExternalTenants(clusterProperties.TrustedExternalTenants))
 		d.Set("enable_disk_encryption", clusterProperties.EnableDiskEncryption)
 		d.Set("enable_streaming_ingest", clusterProperties.EnableStreamingIngest)
 		d.Set("enable_purge", clusterProperties.EnablePurge)
